requirements.txt
Django==2.0
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0boto3==1.13.11
botocore==1.16.11
Django==2.0
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0
django-storages==1.9.1
docutils==0.15.2
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
boto3==1.13.11
botocore==1.16.11
dj-database-url==0.5.0
Django==2.0
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
whitenoise==5.0.1
